blog done with local host strapi
