+++ frontend/next.config.mjs
@@ -9,6 +9,12 @@ const nextConfig = {
                 port: '',
                 pathname: '/**', // all paths under this hostname are allowed
             },
+            {
+                protocol: 'http',
+                hostname: 'localhost',
+                port: '1337',
+                pathname: '/**', // all paths under this hostname are allowed
+            }
         ],
     },
     optimizeFonts: false,
