+++ frontend/src/api/blog/blog.api.ts
@@ -1,35 +1,5 @@
 import request from '../../utils/request';
-
-// Interface matching your Strapi content type structure
-export interface BlogPost {
-  id: number;
-  attributes: {
-    Title: string;
-    Slug: string;
-    Content: string;
-    Author: string;
-    Published: boolean;
-    CoverImage: {
-      data: {
-        id: number;
-        attributes: {
-          url: string;
-          width: number;
-          height: number;
-          formats: {
-            thumbnail: { url: string };
-            small: { url: string };
-            medium: { url: string };
-            large?: { url: string };
-          };
-        };
-      } | null;
-    };
-    createdAt: string;
-    updatedAt: string;
-    publishedAt: string;
-  };
-}
+import { BlogPost } from '../../components/blog/BlogCard';
 
 export interface BlogResponse {
   data: BlogPost[];
@@ -45,37 +15,28 @@ export interface BlogResponse {
 
 export const fetchBlogPosts = async (page = 1, pageSize = 10) => {
   try {
-    console.log(`Fetching blog posts from http://localhost:1337/api/posts`);
-    console.log(`With params: page=${page}, pageSize=${pageSize}`);
-    
     const response = await request.get('http://localhost:1337/api/posts', {
       params: {
         'pagination[page]': page,
         'pagination[pageSize]': pageSize,
-        'populate': 'CoverImage',
         'sort': 'publishedAt:desc',
         'filters[Published][$eq]': true
       }
     });
     
-    console.log('Full API response:', response);
-    console.log('Response data structure:', JSON.stringify(response.data, null, 2));
-    
     return response.data;
-  } catch (error: any) {
+  } catch (error) {
     console.error('Failed to fetch blog posts:', error);
-    console.error('Error response:', error.response?.data);
     throw error;
   }
 };
-
 
 export const fetchBlogPostBySlug = async (slug: string) => {
   try {
     const response = await request.get('http://localhost:1337/api/posts', {
       params: {
         'filters[Slug][$eq]': slug,
-        'populate': 'CoverImage' // Include the cover image
+         'populate': 'CoverImage'
       }
     });
