--- frontend/src/app/blog/layout.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export default function BlogLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="min-h-screen bg-gray-50 py-8">
-      {children}
-    </div>
-  );
-}
